feat: 1 elligator + order re-optimized + only one point scalar mult in length 1 AB
CTIDH GA is optimized, better than the original version from CTIDH's source code.
We believe it is secure (heuristically, the joint distribution of elligator's output points' orders should be nearly random).
The effect is small, but if we use Velusqrt or even consider parallel isogeny, it would be more useful.

Signed-off-by: Yifan Zheng <su1yu4n@gmail.com>